Cell and CellCube scripts, MazeGeneration Wip
2 1/2 Hours
Cell and CellCube script:  i added the cellCube scripts to deactive the walls of the maze and the Cell script that is the parent of the CellCube, because i wanted to add triangle or hexagon mazes later on.
generation algorithm: i added methods to start the generation, for the recursive visiting of cells,
choosing of a random neighbour.
Its still WIP, checking the stack for new neighbours is next.